Make XML saving/loading work in Java 8.
HwComponentModel used a strange setup in order to save the list of interfaces
using XML.  Strangely, this worked in Java 7 for saving/loading, but in Java 8
it worked only for loading. This commit removes the (get|set)InterfacesAsList()
operations from HwComponentModel and thus saves the interfaces in an
<interfacesMap> tag, as the other maps. This means, of course, that
pre-existing XML network descriptions can no longer be used. The examples
have been replaced accordingly.